refactor(ouroborus): args in a proper file
