+++ ouroborus/test/test_container.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../lib/ouroborus/container'
+require_relative '../lib/ouroborus/args'
 
 require 'rspec/autorun'
